refactor(errors): reduce shared error surface ownership
