+++ src/frame_compare/errors.py
@@ -1,19 +1,8 @@
-"""Frame Compare error types (minimal stub for config module).
-
-Error Codes:
-    FC-1001: CONFIG_NOT_FOUND
-    FC-1002: CONFIG_PARSE_ERROR
-    FC-1003: CONFIG_VALIDATION_ERROR
-    FC-1004: PRESET_NOT_FOUND
-    FC-1005: PRESET_INVALID
-    FC-1006: PRESET_NAME_INVALID
-    FC-1007: CONFIG_WRITE_ERROR
-"""
+"""Frame Compare base error types and common core exceptions."""
 
 from __future__ import annotations
 
 from dataclasses import dataclass
-from enum import IntEnum
 from pathlib import Path
 from typing import TYPE_CHECKING, cast
 
@@ -107,198 +96,11 @@ class FrameCompareError(Exception):
         return f"{self.__class__.__name__}({self.context!r})"
 
 
-class ConfigError(FrameCompareError):
-    """Base class for configuration errors."""
-
-
-class ConfigNotFoundError(ConfigError):
-    """Configuration file not found (FC-1001)."""
-
-    def __init__(self, path: Path) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-1001",
-                name="CONFIG_NOT_FOUND",
-                message=f"Configuration file not found: {path}",
-                hint="Run 'frame-compare wizard' or create config/config.toml",
-                details={"path": str(path)},
-            )
-        )
-        self.path = path
-
-
-class ConfigParseError(ConfigError):
-    """TOML parsing failed (FC-1002)."""
-
-    def __init__(self, path: Path, parse_details: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-1002",
-                name="CONFIG_PARSE_ERROR",
-                message=f"Failed to parse {path}: {parse_details}",
-                hint="Check TOML syntax at the indicated line",
-                details={"path": str(path), "parse_error": parse_details},
-            )
-        )
-        self.path = path
-
-
-class ConfigValidationError(ConfigError):
-    """Config validation failed (FC-1003)."""
-
-    def __init__(self, errors: list[dict[str, JSONValue]]) -> None:
-        fields: list[str] = []
-        for e in errors:
-            loc = e.get("loc")
-            if isinstance(loc, list) and loc:
-                fields.append(str(loc[-1]))
-            else:
-                fields.append("unknown")
-
-        # Cast to avoid invariance issues with list[dict[str, JSONValue]] vs list[JSONValue]
-        safe_errors = cast("JSONValue", errors)
-
-        super().__init__(
-            ErrorContext(
-                code="FC-1003",
-                name="CONFIG_VALIDATION_ERROR",
-                message=f"Invalid configuration: {', '.join(fields)}",
-                hint="Check field types and constraints",
-                details={"validation_errors": safe_errors},
-            )
-        )
-        self.validation_errors = errors
-
-
-class PresetNotFoundError(ConfigError):
-    """Preset not found (FC-1004)."""
-
-    def __init__(self, name: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-1004",
-                name="PRESET_NOT_FOUND",
-                message=f"Preset not found: {name}",
-                hint="Run 'frame-compare preset list' to see available",
-                details={"preset_name": name},
-            )
-        )
-        self.preset_name = name
-
-
-class PresetInvalidError(ConfigError):
-    """Preset file has invalid TOML syntax (FC-1005)."""
-
-    def __init__(self, path: Path, parse_details: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-1005",
-                name="PRESET_INVALID",
-                message=f"Invalid preset file: {path}",
-                hint="Check TOML syntax in preset file",
-                details={"path": str(path), "parse_error": parse_details},
-            )
-        )
-        self.path = path
-
-
-class PresetNameInvalidError(ConfigError):
-    """Preset name is invalid (FC-1006)."""
-
-    def __init__(self, name: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-1006",
-                name="PRESET_NAME_INVALID",
-                message=f"Invalid preset name: {name}",
-                hint="Use only letters, numbers, '_' and '-'",
-                details={"preset_name": name},
-            )
-        )
-        self.preset_name = name
-
-
-class ConfigWriteError(ConfigError):
-    """Configuration or preset write failed (FC-1007)."""
-
-    def __init__(self, path: Path, *, label: str, cause: OSError) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-1007",
-                name="CONFIG_WRITE_ERROR",
-                message=f"Failed to write {label}: {path}",
-                hint="Check that the destination path is writable",
-                details={"path": str(path), "error": str(cause)},
-                cause=cause,
-            )
-        )
-        self.path = path
-
-
 # ─── 3.2 Dependency Errors (FC-2xxx) ───────────────────────────────────────────
 
 
 class DependencyError(FrameCompareError):
     """Base class for dependency failures (VapourSynth, FFmpeg, plugins)."""
-
-
-class VapourSynthNotFoundError(DependencyError):
-    """VapourSynth module not found (FC-2001)."""
-
-    def __init__(self) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-2001",
-                name="VAPOURSYNTH_NOT_FOUND",
-                message="VapourSynth python module not found",
-                hint="Install VapourSynth (pip install VapourSynth) or check PYTHONPATH",
-            )
-        )
-
-
-class VapourSynthError(DependencyError):
-    """VapourSynth core error (FC-2002)."""
-
-    def __init__(self, details: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-2002",
-                name="VAPOURSYNTH_ERROR",
-                message=f"VapourSynth core error: {details}",
-                hint="Check script syntax or plugin compatibility",
-                details={"vs_error": details},
-            )
-        )
-
-
-class PluginNotFoundError(DependencyError):
-    """VapourSynth plugin missing (FC-2003)."""
-
-    def __init__(self, namespace: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-2003",
-                name="PLUGIN_NOT_FOUND",
-                message=f"Required VapourSynth plugin missing: {namespace}",
-                hint=f"Install plugin providing namespace '{namespace}'",
-                details={"namespace": namespace},
-            )
-        )
-
-
-class LibplaceboError(DependencyError):
-    """vs-placebo specific error (FC-2004)."""
-
-    def __init__(self, details: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-2004",
-                name="LIBPLACEBO_ERROR",
-                message=f"Libplacebo error: {details}",
-                hint="Check Vulkan drivers or shader support",
-                details={"libplacebo_error": details},
-            )
-        )
 
 
 class FFmpegNotFoundError(DependencyError):
@@ -354,20 +156,6 @@ class VSPreviewNotFoundError(DependencyError):
                 name="VSPREVIEW_NOT_FOUND",
                 message="VSPreview not installed",
                 hint="Install vspreview (and a Qt backend) for interactive alignment verification",
-            )
-        )
-
-
-class TonemapRequiresVapourSynthError(DependencyError):
-    """HDR tonemapping requires VapourSynth rendering (FC-2009)."""
-
-    def __init__(self) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-2009",
-                name="TONEMAP_REQUIRES_VAPOURSYNTH",
-                message="HDR tonemapping requires VapourSynth rendering",
-                hint="Disable [color].enable_tonemap for FFmpeg-only extraction, or use VapourSynth",
             )
         )
 
@@ -830,130 +618,6 @@ class NetworkError(FrameCompareError):
     """Base class for network failures."""
 
 
-class NetworkUnreachableError(NetworkError):
-    """No internet connection (FC-5001)."""
-
-    def __init__(self) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-5001",
-                name="NETWORK_UNREACHABLE",
-                message="Network unreachable",
-                hint="Check internet connection",
-            )
-        )
-
-
-class SlowpicsError(NetworkError):
-    """General slow.pics API failure (FC-5002)."""
-
-    def __init__(self, reason: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-5002",
-                name="SLOWPICS_ERROR",
-                message=f"slow.pics error: {reason}",
-                hint="Check service status",
-                details={"reason": reason},
-            )
-        )
-
-
-class SlowpicsRateLimitedError(NetworkError):
-    """Too many requests to slow.pics (FC-5003)."""
-
-    def __init__(self) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-5003",
-                name="SLOWPICS_RATE_LIMITED",
-                message="slow.pics rate limit exceeded",
-                hint="Wait before retrying",
-            )
-        )
-
-
-class SlowpicsUnavailableError(NetworkError):
-    """slow.pics service down (FC-5004)."""
-
-    def __init__(self) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-5004",
-                name="SLOWPICS_UNAVAILABLE",
-                message="slow.pics service unavailable",
-                hint="Try again later",
-            )
-        )
-
-
-class TmdbError(NetworkError):
-    """TMDB API failure (FC-5005)."""
-
-    def __init__(self, reason: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-5005",
-                name="TMDB_ERROR",
-                message=f"TMDB error: {reason}",
-                hint="Check API key",
-                details={"reason": reason},
-            )
-        )
-
-
-class TmdbRateLimitedError(NetworkError):
-    """Too many requests to TMDB (FC-5006)."""
-
-    def __init__(self) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-5006",
-                name="TMDB_RATE_LIMITED",
-                message="TMDB rate limit exceeded",
-                hint="Wait before retrying",
-            )
-        )
-
-
-class NetworkTimeoutError(NetworkError):
-    """Request timed out (FC-5007)."""
-
-    def __init__(self, url: str, timeout: float) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-5007",
-                name="NETWORK_TIMEOUT",
-                message=f"Request to {url} timed out after {timeout}s",
-                hint="Check connection speed",
-                details={"url": url, "timeout": timeout},
-            )
-        )
-
-
-class SSLError(NetworkError):
-    """Certificate verification failed (FC-5008)."""
-
-    def __init__(self, reason: str) -> None:
-        super().__init__(
-            ErrorContext(
-                code="FC-5008",
-                name="SSL_ERROR",
-                message=f"SSL verification failed: {reason}",
-                hint="Check system certificates",
-                details={"reason": reason},
-            )
-        )
-
-
-class ServiceError(NetworkError):
-    """Marker base for service failures."""
-
-
-class PublishError(NetworkError):
-    """Marker base for publishing failures."""
-
-
 # ─── 3.6 Internal Errors (FC-9xxx) ─────────────────────────────────────────────
 
 
@@ -1004,63 +668,3 @@ class UnexpectedStateError(InternalError):
                 details={"state": details},
             )
         )
-
-
-# ─── 4. Exit Code Mapping ──────────────────────────────────────────────────────
-
-
-class ExitCode(IntEnum):
-    SUCCESS = 0
-    GENERAL_ERROR = 1
-    CONFIG_ERROR = 2
-    DEPENDENCY_ERROR = 3
-    INPUT_ERROR = 4
-    PROCESSING_ERROR = 5
-    NETWORK_ERROR = 6
-    INTERRUPTED = 130
-
-
-def get_exit_code(error: FrameCompareError) -> ExitCode:
-    """Map exception to exit code."""
-    code = error.code
-    if code.startswith("FC-"):
-        category = code.split("-", 1)[1][:1]
-        if category == "1":
-            return ExitCode.CONFIG_ERROR
-        if category == "2":
-            return ExitCode.DEPENDENCY_ERROR
-        if category == "3":
-            return ExitCode.INPUT_ERROR
-        if category == "4":
-            return ExitCode.PROCESSING_ERROR
-        if category == "5":
-            return ExitCode.NETWORK_ERROR
-    # InternalError and unknown FrameCompareErrors map to 1
-    return ExitCode.GENERAL_ERROR
-
-
-# ─── 5. Error Formatting Utilities ─────────────────────────────────────────────
-
-
-def format_error_console(error: FrameCompareError, *, verbose: bool = False) -> str:
-    """Format error for console output."""
-    output = f"✗ Error [{error.code}]: {error.context.message}\n"
-    if error.hint:
-        output += f"  Hint: {error.hint}\n"
-
-    if verbose and error.context.details:
-        output += "\n  Details:\n"
-        for k, v in error.context.details.items():
-            output += f"    {k}: {v}\n"
-    elif not verbose and error.context.details:
-        output += "\n  For more details, run with --verbose"
-
-    return output.rstrip()
-
-
-def format_error_json(error: FrameCompareError) -> dict[str, JSONValue]:
-    """Format error for JSON output."""
-    return {
-        "success": False,
-        "error": error.context.to_dict(),
-    }
