+++ src/frame_compare/orchestration/preflight.py
@@ -10,10 +10,10 @@ import os
 from dataclasses import dataclass, field
 from pathlib import Path
 
+from frame_compare.config.errors import ConfigNotFoundError
 from frame_compare.config.loader import load_config
 from frame_compare.config.schema import ConfigSchema
 from frame_compare.errors import (
-    ConfigNotFoundError,
     DirectoryNotFoundError,
     InputDiscoveryError,
     NoVideosFoundError,
