+++ tests/analysis/test_metrics.py
@@ -43,9 +43,9 @@ from frame_compare.analysis.types import FrameMetrics  # noqa: E402
 from frame_compare.config.schema import AnalysisConfig  # noqa: E402
 from frame_compare.errors import (  # noqa: E402
     MetricsCalculationError,
-    PluginNotFoundError,
     SourceLoadError,
 )
+from frame_compare.vs.errors import PluginNotFoundError  # noqa: E402
 
 
 class MockFrame:
