+++ tests/integration/test_loadsources_probe_cache.py
@@ -7,10 +7,10 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from frame_compare.errors import VapourSynthError, VapourSynthNotFoundError
 from frame_compare.orchestration.coordinator import RunDependencies, RunRequest, execute_run
 from frame_compare.orchestration.probing import load_clip_probe_cache
 from frame_compare.vs.env import detect_plugins, ensure_vs_environment
+from frame_compare.vs.errors import VapourSynthError, VapourSynthNotFoundError
 from frame_compare.vs.loader import DefaultVSLoader, VSLoader
 from frame_compare.vs.types import SourceInfo
