+++ tests/orchestration/test_execute_run.py
@@ -14,16 +14,14 @@ import pytest
 from PIL import Image
 
 import frame_compare.analysis.cache_io as cache_io
+from frame_compare.config.errors import ConfigNotFoundError
 from frame_compare.config.loader import load_config
 from frame_compare.config.schema import ConfigSchema, OverlayMode, TonemapPreset
 from frame_compare.errors import (
     AudioAlignmentError,
     CacheCorruptionError,
     CacheVersionMismatchError,
-    ConfigNotFoundError,
     MetricsCalculationError,
-    TmdbError,
-    TonemapRequiresVapourSynthError,
 )
 from frame_compare.orchestration import coordinator, phase_tasks, preparation
 from frame_compare.orchestration.context import (
@@ -36,9 +34,11 @@ from frame_compare.orchestration.coordinator import RunDependencies, RunRequest,
 from frame_compare.orchestration.execution import build_execution_phase_plan
 from frame_compare.orchestration.types import ExecutionState, PrepState, RunArtifacts
 from frame_compare.services.alignment import CACHE_FILE_NAME
+from frame_compare.services.errors import TmdbError
 from frame_compare.services.run_folder import derive_run_folder_name
 from frame_compare.services.types import AlignmentResult, MetadataConfig, TmdbMetadata
 from frame_compare.utils.types import WorkspacePaths
+from frame_compare.vs.errors import TonemapRequiresVapourSynthError
 from frame_compare.vs.types import HDRMetadata, SourceInfo
 
 # Minimal valid TOML config content
