+++ tests/render/test_orchestrator.py
@@ -7,11 +7,8 @@ from structlog.testing import capture_logs
 
 from frame_compare.config.schema import ColorConfig, ConfigSchema
 from frame_compare.errors import (
-    PluginNotFoundError,
     RenderError,
     SourceLoadError,
-    TonemapRequiresVapourSynthError,
-    VapourSynthNotFoundError,
 )
 from frame_compare.render.batch.orchestrator import (
     ProgressReporter,
@@ -24,6 +21,11 @@ from frame_compare.render.types import (
     RenderRequest,
     ScreenshotBatchRequest,
     ScreenshotRenderOptions,
+)
+from frame_compare.vs.errors import (
+    PluginNotFoundError,
+    TonemapRequiresVapourSynthError,
+    VapourSynthNotFoundError,
 )
