+++ tests/render/test_tonemap_wiring.py
@@ -15,8 +15,6 @@ from frame_compare.config.overrides import TonemapCliOverrides
 from frame_compare.config.schema import ColorConfig, ConfigSchema, ToneCurve, TonemapPreset
 from frame_compare.errors import (
     SourceLoadError,
-    TonemapRequiresVapourSynthError,
-    VapourSynthNotFoundError,
 )
 from frame_compare.render.batch.orchestrator import render_screenshots
 from frame_compare.render.prepare import (
@@ -24,6 +22,10 @@ from frame_compare.render.prepare import (
     should_tonemap,
 )
 from frame_compare.render.types import ScreenshotRenderOptions
+from frame_compare.vs.errors import (
+    TonemapRequiresVapourSynthError,
+    VapourSynthNotFoundError,
+)
 from frame_compare.vs.types import SourceInfo, TonemapSettings
 
 # ─── Helper Truth Table Tests ──────────────────────────────────────────────────
