+++ tests/services/test_publishers.py
@@ -8,7 +8,7 @@ import httpx
 import pytest
 
 from frame_compare.config.schema import SlowpicsConfig, Visibility
-from frame_compare.errors import (
+from frame_compare.services.errors import (
     SlowpicsError,
     SlowpicsRateLimitedError,
     SlowpicsUnavailableError,
